api auth
Authentication using token and basic authentication
