+++ api_v1/models/__init__.py
@@ -1,0 +1,1 @@
+from .bearerAuthentication import BearerAuthentication
