+++ api_v1/models/bearerAuthentication.py
@@ -1,0 +1,13 @@
+from rest_framework.authentication import TokenAuthentication
+
+
+class BearerAuthentication(TokenAuthentication):
+    '''
+    Simple token based authentication using api-token-auth.
+
+    Clients should authenticate by passing the token key in the 'Authorization'
+    HTTP header, prepended with the string 'Bearer '.  For example:
+
+    Authorization: Bearer 956e252a-513c-48c5-92dd-bfddc364e812
+    '''
+    keyword = 'Bearer'
