+++ api_v1/serializers/topicSerializer.py
@@ -25,6 +25,4 @@ class TopicSerializer(serializers.HyperlinkedModelSerializer):
                 'lookup_field': 'pk'
             },
         }
-        
-        ordering = ['id']
