+++ api_v1/views/__init__.py
@@ -1,8 +1,8 @@
-from .userViewSet import SubscriberViewSet
-from .userViewSet import AuthorViewSet
-from .userViewSet import PublisherViewSet
-from .userViewSet import AdminViewSet
+from .subscriberViewSet import SubscriberViewSet
+from .authorViewSet import AuthorViewSet
+from .publisherViewSet import PublisherViewSet
+from .adminViewSet import AdminViewSet
 from .userViewSet import UserViewSet
 from .groupViewSet import GroupViewSet
 from .articleViewSet import ArticleViewSet
-from .articleViewSet import TopicViewSet
+from .topicViewSet import TopicViewSet
